修改板子参数，适配原子灯，在200-505段积分15次，积分时间500ms，得到的结果比较合适

+++ R_PINGCHANG/Core/Inc/cJSON_Test.h
@@ -7,12 +7,11 @@
 #define H750_CJSON_CJSON_TEST_H
 struct paramstruct {
     char * command;
-    uint8_t Set_Integration_Time;//积分时间最大100ms
+    uint16_t Set_Integration_Time;//积分时间最大100ms
     uint8_t Average_Number;
     int pixel_segment_1;
     int pixel_segment_2;  //segment_2最大为2048
     uint8_t mul_int_max;
-    //char * dma_state;
 };
 
 void cjson(uint8_t g_usart_json_string[]);
